moved validator to separate endpoint
